Add credentials in server-side rendering
Upon first page load, fetch the user sign-in status from the API
server via UI server with the help of the static function at Page
component. Add the cookie header in the fetch request.

In the request made by the browser, the cookie is added in the header
by the browser. But in the request made by the server, the cookie has
to be added progamatically.